Complete_Agentic_AI_Bootcamp_with_LangGraph_and_LangChain/Section_6_Open_AI_and_Ollama/requirements.txt
ipykernel
python-dotenv
langchain
langchain-classic
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
fitz
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
streamlit
